pyproject.toml
[tool.poetry]
name = "project-chess"
version = "0.1.0"
description = ""
authors = ["AkulaFish <denisvoytehovich3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
isort = "^5.13.2"
black = "^24.4.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
